pipeline/requirements.txt
# Python 3.11+
cryptography>=42.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-docx>=1.1.0
pillow>=10.3.0
anthropic>=0.28.0
pytest>=8.2.0
httpx>=0.27.0        # FastAPI test client
python-multipart>=0.0.9
